requirements.txt
streamlit==1.32.0
openai==1.17.0
python-dotenv==1.0.1
httpx==0.27.0
streamlit-audiorecorder==0.0.6
transformers==4.44.0
torch>=2.0.0
librosa==0.10.1
numpy>=1.24.0
pydub>=0.25.1
scipy>=1.10.0
soundfile>=0.12.1
datasets>=2.14.0
accelerate>=0.20.0
typing-extensions>=4.5.0
